+++ app/Http/Resources/Api/V2/ItemResource.php
@@ -18,7 +18,9 @@ class ItemResource extends JsonResource
             'id' => $this->id,
             'title' => $this->title,
             'authors' => $this->getUniqueAuthorsWithAuthorityNames(),
-            'dating' => $this->dating,
+            'dating' => $this->getDatingFormated(true),
+            'date_earliest' => $this->date_earliest,
+            'date_latest' => $this->date_latest,
             'description' => $this->description,
             'authorities' => AuthorityResource::collection($this->authorities),
         ];