+++ resources/js/app.js
@@ -22,8 +22,16 @@ require('./components/newsletter-signup-form-tracker')
 
 // Vue components
 import Vue from 'vue'
+import VueRouter from 'vue-router'
 import 'livewire-vue'
 import { Lang } from 'laravel-vue-lang';
+
+Vue.use(VueRouter)
+
+const router = new VueRouter({
+    mode: 'history',
+});
+
 
 
 window.Vue = Vue
@@ -32,8 +40,13 @@ Vue.component('featured-piece-click-tracker', require('./components/FeaturedPiec
 Vue.component('filter-sort-by', require('./components/filter/SortBy.vue').default);
 Vue.component('filter-checkbox', require('./components/filter/Checkbox.vue').default);
 Vue.component('filter-custom-select', require('./components/filter/CustomSelect.vue').default);
+Vue.component('filter-new', require('./components/catalog-new/Filter.vue').default);
 Vue.component('filter-new-custom-select', require('./components/catalog-new/NewCustomSelect.vue').default);
+Vue.component('filter-new-custom-checkbox', require('./components/catalog-new/NewCustomCheckbox.vue').default);
+Vue.component('filter-new-selected-labels', require('./components/catalog-new/NewSelectedLabels.vue').default);
+Vue.component('filter-new-sort', require('./components/catalog-new/NewSort.vue').default);
 Vue.component('filter-new-mobile-custom-select', require('./components/catalog-new/NewMobileCustomSelect.vue').default);
+Vue.component('filter-show-more', require('./components/catalog-new/ShowMore.vue').default);
 Vue.component('flickity', require('./components/Flickity.vue').default);
 Vue.component('home.shuffle-orchestrator', require('./components/home/ShuffleOrchestrator.vue').default);
 Vue.component('home.transition-in-place', require('./components/home/TransitionInPlace.vue').default);
@@ -64,4 +77,4 @@ Vue.component('color-slider', require('./components/vue/color-slider').default);
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
- new Vue({ el: '#app' });
+ new Vue({ el: '#app', router });